test(eventhub): the M2c test file goes, the M2d test stays
Completes the M2c drop the merge began. eventhub_keysync_test.go was
written for this branch's key-sync fix, which master's 94b88765 replaced;
the only test in it that still covers something master lacks is the
uncomparable-processor refusal, and that moves to a file whose name says
what it is.

The rename is not cosmetic. eventhub_keysync_test.go would have described
a mechanism the tree no longer has, and a test file named after a deleted
fix is how a reader ends up looking for one.

Both files were left uncommitted by the merge — the deletion was staged
through a plain rm and the new file was never added — so the merge commit
still tracked the old one while the working tree had neither. That is why
the local build and tests passed over a state git did not hold.
